Revert "DolphinQt: use default dpi rounding mode (passthrough)"
This reverts commit cfe36836688ae21ae32383958504fbc5f6c21660.

On Windows systems with 125% DPI scaling, this was causing both icons
and the OSD to be upscaled by 25% using nearest neighbor scaling, which
looks really bad. shuffle2 has said that this will be improved later,
but we should revert it for now for the sake of the upcoming beta build.